Change chord results to include base interval.
Previously a chord recognition result included the note name of the lowest
note. This was to allow for the representation of slash chords.

This change instead stores the interval between the root of the chord
and the lowest note.

This keeps things a bit more generic and allows for smarter notation of
chord intervals at somepoint.